Add basic dictation engine
Add a dictation engine that pauses whenever it's exceeded the word rate.
This is okay, but it's not great, since it tends to sprint and pause
frequently, rather than pace itsel like a human dictator.

The next step is to make the dictation engine try to say the words on a
"schedule", rather than sprint and pause repeatedly.